# Speed up Jest with changes from template project