--- a/.husky/pre-commit
+++ b/.husky/pre-commit
@@ -1,4 +1,4 @@
 #!/bin/sh
 . "$(dirname "$0")/_/husky.sh"
 
-NODE_ENV=dev && node_modules/.bin/lint-staged && node_modules/.bin/tsc && npm test
+NODE_ENV=dev && node_modules/.bin/lint-staged && npm run typecheck && npm test