--- a/package.json
+++ b/package.json
@@ -21,7 +21,8 @@
     "record-build-info": "node ./bin/record-build-info",
     "lint": "eslint . --cache --max-warnings 0",
     "typecheck": "tsc && tsc -p integration_tests",
-    "test": "jest --runInBand --detectOpenHandles",
+    "test": "jest --coverage",
+    "test:ci": "jest --runInBand --coverage",
     "security_audit": "npx audit-ci --config audit-ci.json",
     "int-test": "cypress run --config video=false --env $(grep ONE_TIME_CODE_AUTH_ENABLED feature.env)",
     "int-test-ui": "cypress open --env $(grep ONE_TIME_CODE_AUTH_ENABLED feature.env)",
@@ -33,7 +34,14 @@
   },
   "jest": {
     "preset": "ts-jest",
-    "collectCoverage": true,
+    "transform": {
+      "^.+\\.tsx?$": [
+        "ts-jest",
+        {
+          "isolatedModules": true
+        }
+      ]
+    },
     "collectCoverageFrom": [
       "server/**/*.{ts,js,jsx,mjs}"
     ],
@@ -63,8 +71,7 @@
       "json",
       "node",
       "ts"
-    ],
-    "testRunner": "jest-jasmine2"
+    ]
   },
   "nodemonConfig": {
     "ignore": [
@@ -166,7 +173,6 @@
     "jest": "^29.3.1",
     "jest-environment-jsdom": "^29.3.1",
     "jest-html-reporter": "^3.7.0",
-    "jest-jasmine2": "^29.2.1",
     "jest-junit": "^14.0.1",
     "lint-staged": "^13.0.3",
     "mocha-junit-reporter": "^2.2.0",