--- a/server/data/healthCheck.test.ts
+++ b/server/data/healthCheck.test.ts
@@ -15,6 +15,7 @@ describe('Service healthcheck', () => {
   })
 
   afterEach(() => {
+    nock.abortPendingRequests()
     nock.cleanAll()
   })