* Bring over recent changes from the template project to speed up Jest.
* Remove custom test runner, `jest-jasmine2` - which looks to have been added early in the project to [work around a bug](https://github.com/ministryofjustice/send-legal-mail-to-prisons/pull/222#discussion_r821771900) but is no longer needed.
* Tweak a few tests to prevent warnings about open handles being left on test runs

Performance improvements (running locally):

**Before**
```bash
Test Suites: 109 passed, 109 total
Tests:       598 passed, 598 total
Snapshots:   0 total
Time:        28.109 s, estimated 35 s
```

**After**
```bash
Test Suites: 109 passed, 109 total
Tests:       598 passed, 598 total
Snapshots:   0 total
Time:        7.582 s, estimated 14 s
```